Update and rename README.md to The Evolution of Internet Explorer
Sure, here's an article about the versions of Internet Explorer:

# The Evolution of Internet Explorer

Internet Explorer, often abbreviated as IE, was a series of graphical web browsers developed by Microsoft and included as part of the Microsoft Windows line of operating systems.

## Internet Explorer 1.0

Internet Explorer 1.0 was first released as part of the add-on package Plus! for Windows 95 in August 1995. It was a reworked version of Spyglass Mosaic which Microsoft had licensed.

## Internet Explorer 3.0

Internet Explorer 3.0, released in 1996, was the first version to provide a full browsing experience, with support for background sounds, images, tables, and other HTML enhancements.

## Internet Explorer 5.0

Internet Explorer 5.0, released in 1999, was the first version to integrate more closely with Windows, allowing users to browse the web directly from the Windows desktop.

## Internet Explorer 6.0

Internet Explorer 6.0, released in 2001, was the most widely used web browser during its tenure (surpassing Netscape). It was the last version to be based on the Trident layout engine.

## Internet Explorer 7.0

Internet Explorer 7.0, released in 2006, was the first version to include tabbed browsing, a highly requested feature at the time.

## Internet Explorer 8.0

Internet Explorer 8.0, released in 2009, added several new features, including a search box and improved security.

## Internet Explorer 9.0

Internet Explorer 9.0, released in 2011, was the first version to support HTML5 and CSS3.

## Internet Explorer 10.0

Internet Explorer 10.0, released in 2012, was the first version to be included with Windows 8.

## Internet Explorer 11.0

Internet Explorer 11.0, released in 2013, is the final version of Internet Explorer. It was officially discontinued in June 2022.

In conclusion, Internet Explorer has had a long and storied history, from its inception in 1995 to its final version in 2013. Despite being discontinued, it played a significant role in shaping the internet as we know it today.